Restrict user from unusable chest inventory slots
